Add padlock icon next to any link leading to private resources
I like the idea introduced in https://github.com/artsy/README/pull/78 to
add a padlock icon next to external links leading to content unavailable
to readers outside of the Artsy organization.

I've added that helpful icon to other links in the repository. Probably
missed a bunch, but I think this covers most.