docs: Add Why's Poignant Guide to Ruby
I still think Why's Guide is an excellent tool for communicating a lot
of context regarding the Ruby community and *mentality*. There are a
number of incredibly helpful (and entirely absurd) examples from _why
that taught me a lot about the flexibility and expressiveness of Ruby as
a programming language.

Plus, I think the guide qualifies as art and this is Artsy.